% $Id: is.FLBREM.Rd,v 1.6.2.1 2005/12/19 11:13:48 lkell Exp $
\name{is.FLBREM}
\alias{is.FLBREM}
\title{FLBREM}
\description{Testing of FLBREM objects}
\usage{
  is.FLBREM(x)
}
\arguments{
  \item{x}          {object to test.}
}
\details{
}
\value{
	\code{is.FLBREM} returns \code{TRUE} or \code{FALSE}
}
\seealso{
    \code{\link{FLBREM-class}}
}
\examples{{
    is.FLBREM(new("FLBREM"))
}
}
\author{Laurence Kell, Verena Trenkle}
\keyword{classes}
\keyword{manip}
